Página Cuenta mejoras

+++ database/migrations/2025_05_01_174059_modify_dietas_table.php
@@ -9,21 +9,32 @@ class ModifyDietasTable extends Migration
     public function up()
     {
         Schema::table('dietas', function (Blueprint $table) {
-            $table->unsignedBigInteger('id_usuario')->after('id');
-            $table->binary('pdf_dieta')->nullable()->after('id_usuario');
-            $table->date('fecha_generacion')->after('pdf_dieta');
-            $table->unsignedBigInteger('id_mascota')->nullable()->change();
-
-            $table->foreign('id_usuario')->references('id')->on('usuarios')->onDelete('cascade');
+            if (!Schema::hasColumn('dietas', 'id_usuario')) {
+                $table->unsignedBigInteger('id_usuario')->after('id');
+                $table->foreign('id_usuario')->references('id')->on('usuarios')->onDelete('cascade');
+            }
+            if (!Schema::hasColumn('dietas', 'pdf_dieta')) {
+                $table->binary('pdf_dieta')->nullable()->after('id_usuario');
+            }
+            if (!Schema::hasColumn('dietas', 'fecha_generacion')) {
+                $table->date('fecha_generacion')->after('pdf_dieta');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('dietas', function (Blueprint $table) {
-            $table->dropForeign(['id_usuario']);
-            $table->dropColumn(['id_usuario', 'pdf_dieta', 'fecha_generacion']);
-            $table->unsignedBigInteger('id_mascota')->nullable(false)->change();
+            if (Schema::hasColumn('dietas', 'id_usuario')) {
+                $table->dropForeign(['id_usuario']);
+                $table->dropColumn('id_usuario');
+            }
+            if (Schema::hasColumn('dietas', 'pdf_dieta')) {
+                $table->dropColumn('pdf_dieta');
+            }
+            if (Schema::hasColumn('dietas', 'fecha_generacion')) {
+                $table->dropColumn('fecha_generacion');
+            }
         });
     }
 }
